requirements.txt
ollama
aiohttp
sounddevice
numpy
requests
websockets
faster-whisper
soundfile
keyboard
pyyaml
openai
PyQt6